@@ -1,40 +1,36 @@
 import React, { Component } from 'react';
+import { withStyles } from '@material-ui/core/styles';
+import Paper from '@material-ui/core/Paper';
+import Grid from '@material-ui/core/Grid';
+import { MuiThemeProvider, createMuiTheme } from '@material-ui/core/styles';
+
 import BooksList from './BooksList';
 import Credit from './Credit';
 import {inject, observer} from "mobx-react";
 import AddBook from "./buttons/AddBook";
+import Header from './Header';
+import BookModal from './BookModal';
+
+const theme = createMuiTheme({
+    overrides: {
+        // Name of the component ⚛️ / style sheet
+        MuiButton: {
+
+        }
+    },
+});
 
 class App extends Component {
-
-    addBook(){
-        this.props.BooksStore.editBook({
-            bookId: "",
-            title: "",
-            date: new Date()
-        });
-    }
-
     render() {
         return (
-            <div id="app-container" className="container-fluid">
-                <div className="row">
-                    <div className="col-sm-12">
-                        <h1 id="app-title">Books Shelf</h1>
-                    </div>
-                </div>
-                <div className="row">
-                    <div className="col-sm-5"></div>
-                    <div className="col-sm-2">
-                        <img id="books-picture" src="images/books.jpg" alt="" />
-                    </div>
-                    <div className="col-sm-5"></div>
-                </div>
-                <AddBook onClick={this.addBook.bind(this)} />
+            <MuiThemeProvider theme={theme}>
+                <Header />
                 <BooksList />
                 <Credit />
-            </div>
+                <BookModal />
+            </MuiThemeProvider>
         );
     }
 }
 
-export default inject("BooksStore")(observer(App));
+export default App;
