@@ -1,9 +1,9 @@
 import React from "react";
-import {Button} from "reactstrap";
+import Button from '@material-ui/core/Button';
 
 export default (props) => {
     return (
-        <Button color="default" onClick={this.addNewBook}>
+        <Button color="default" onClick={props.addNewBook}>
             <span className="mdi mdi-create"></span>
             Add Book
         </Button>
